Adjusted ArrayItem::getMod and IndexSequence::getMod to return long instead of int. Updated CollectionFeatures::pad to handle zeros differently and updated PackedCollectionPad::generate and PackedCollectionPad::delta to use it instead of the constructor directly.

+++ code/src/main/java/io/almostrealism/util/ArrayItem.java
@@ -136,7 +136,7 @@ public class ArrayItem<T> implements Sequence<T> {
 		return values;
 	}
 
-	public int getMod() { return mod; }
+	public long getMod() { return mod; }
 
 	@Override
 	public long lengthLong() { return len; }
